+++ outputs/default.nix
@@ -23,22 +23,7 @@ let
       #   # To use chrome, we need to allow the installation of non-free software
       #   config.allowUnfree = true;
       # };
-      pkgs-2505 = import inputs.nixpkgs-2505 {
-        inherit system;
-        # To use chrome, we need to allow the installation of non-free software
-        config.allowUnfree = true;
-      };
-      pkgs-stable = import inputs.nixpkgs-stable {
-        inherit system;
-        # To use chrome, we need to allow the installation of non-free software
-        config.allowUnfree = true;
-      };
       pkgs-patched = import inputs.nixpkgs-patched {
-        inherit system;
-        # to use chrome, we need to allow the installation of non-free software
-        config.allowUnfree = true;
-      };
-      pkgs-master = import inputs.nixpkgs-master {
         inherit system;
         # to use chrome, we need to allow the installation of non-free software
         config.allowUnfree = true;